Move out classes from main dart file, fix a crash where no content script can be contacted, some localization strings
