+++ test/widget_test.dart
@@ -8,12 +8,12 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:browser_extension/main.dart';
+import 'package:browser_extension/widgets/extension.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(MyExtension());
+    await tester.pumpWidget(Extension());
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
